Input validation with "Celebrate"

+++ backend/src/controllers/IncidentController.js
@@ -48,15 +48,17 @@ module.exports = {
       .select("ong_id")
       .first();
 
-    if (incident.ong_id !== ong_id) {
-      return response.status(401).json({
-        error: "Operation not permitted."
-      });
-    }
+    if (incident) {
+      if (incident.ong_id !== ong_id) {
+        return response.status(401).json({
+          error: "Operation not permitted."
+        });
+      }
 
-    await connection("incidents")
-      .where("id", id)
-      .delete();
+      await connection("incidents")
+        .where("id", id)
+        .delete();
+    }
 
     return response.status(204).send();
   }
